backend/app/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.7.5
anyio==3.5.0; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3; python_version >= '3.6'
cryptography==36.0.1
dnspython==2.2.0; python_version >= '3.6' and python_version < '4'
ecdsa==0.17.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.1.3
fastapi==0.72.0
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0
h11==0.12.0; python_version >= '3.6'
httptools==0.3.0
idna==3.3; python_version >= '3.5'
mako==1.1.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.0.1; python_version >= '3.6'
passlib[bcrypt]==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0; python_full_version >= '3.6.1'
python-dotenv==0.19.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pyyaml==6.0
rsa==4.8; python_version >= '3.6' and python_version < '4'
setuptools==60.5.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy==1.4.29; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
starlette==0.17.1; python_version >= '3.6'
typing-extensions==4.0.1; python_version >= '3.6'
uvicorn[standard]==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
